Adopt "App" namespace
A generic project namespace eliminates the need to think of a cool project identifier and rename various occurrences.

Changed:
- Renamed "Boilerplate" to "App"
- Renamed `AbstractAppTemplate` to `AbstractTemplate`
- Update installation instructions to reflect generic namespace

Removed:
- `Charcoal\Script\RenameScript`
- `Charcoal\Script\KeyNormalizerTrait`
