+++ config/config.sample.json
@@ -1,7 +1,7 @@
 {
     "databases": {
         "default": {
-            "database": "charcoal_boilerplate",
+            "database": "charcoal",
             "username": "root",
             "password": ""
         }
