+++ metadata/app/template/home-template.json
@@ -1,0 +1,8 @@
+{
+    "properties": {
+        "featured_section": {
+            "type": "object",
+            "obj_type": "app/object/section"
+        }
+    }
+}
