--- metadata/boilerplate/object/section.json
@@ -1,13 +1,0 @@
-{
-    "properties": {
-        "master": {
-            "obj_type": "boilerplate/object/section"
-        }
-    },
-    "sources": {
-        "default": {
-            "table": "boilerplate_sections"
-        }
-    },
-    "default_source": "default"
-}
