+++ src/App/Attachment/File.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Attachment;
+
+use Charcoal\Attachment\Object\File as FileAttachment;
+
+/**
+ * Content Block: File
+ */
+class File extends FileAttachment
+{
+}
