+++ src/App/Attachment/Gallery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Attachment;
+
+use Charcoal\Attachment\Object\Gallery as GalleryAttachment;
+
+/**
+ * Content Block: Gallery
+ */
+class Gallery extends GalleryAttachment
+{
+}
