+++ src/App/Attachment/Image.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Attachment;
+
+use Charcoal\Attachment\Object\Image as ImageAttachment;
+
+/**
+ * Content Block: Image
+ */
+class Image extends ImageAttachment
+{
+}
