+++ src/App/Attachment/Text.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Attachment;
+
+use Charcoal\Attachment\Object\Text as TextAttachment;
+
+/**
+ * Content Block: Text
+ */
+class Text extends TextAttachment
+{
+}
