+++ src/App/Attachment/Video.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Attachment;
+
+use Charcoal\Attachment\Object\Video as VideoAttachment;
+
+/**
+ * Content Block: Video
+ */
+class Video extends VideoAttachment
+{
+}
