+++ src/App/Template/GenericTemplate.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Template;
+
+// From App
+use App\Template\AbstractTemplate;
+
+/**
+ * Template Controller: Generic
+ */
+class GenericTemplate extends AbstractTemplate
+{
+}
