--- src/Boilerplate/Attachment/File.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Boilerplate\Attachment;
-
-use Charcoal\Attachment\Object\File as FileAttachment;
-
-/**
- * File attachment
- */
-class File extends FileAttachment
-{
-}
