--- src/Boilerplate/Attachment/Gallery.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Boilerplate\Attachment;
-
-use Charcoal\Attachment\Object\Gallery as GalleryAttachment;
-
-/**
- * Gallery attachment
- */
-class Gallery extends GalleryAttachment
-{
-}
