--- src/Boilerplate/Attachment/Image.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Boilerplate\Attachment;
-
-use Charcoal\Attachment\Object\Image as ImageAttachment;
-
-/**
- * Image attachment
- */
-class Image extends ImageAttachment
-{
-}
