--- src/Boilerplate/Attachment/Text.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Boilerplate\Attachment;
-
-use Charcoal\Attachment\Object\Text as TextAttachment;
-
-/**
- * Text attachment
- */
-class Text extends TextAttachment
-{
-}
