--- src/Boilerplate/Attachment/Video.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Boilerplate\Attachment;
-
-use Charcoal\Attachment\Object\Video as VideoAttachment;
-
-/**
- * Video attachment
- */
-class Video extends VideoAttachment
-{
-}
