--- src/Boilerplate/Template/GenericTemplate.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Boilerplate\Template;
-
-// From 'boilerplate'
-use Boilerplate\Template\AbstractBoilerplateTemplate;
-
-/**
- * Boilerplate Generic Template Controller
- */
-class GenericTemplate extends AbstractBoilerplateTemplate
-{
-}
