+++ views/app/template/home.mustache
@@ -1,0 +1,11 @@
+{{> app/template/partial/site-header }}
+
+<h1>{{ siteName }}</h1>
+
+<h2>Home template.</h2>
+
+<p>Calling the "test" method: <strong>{{ test }}</strong></p>
+
+<p>Template name: {{ templateName }}</p>
+
+{{> app/template/partial/site-footer }}
