--- views/boilerplate/template/home.mustache
@@ -1,11 +1,0 @@
-{{> boilerplate/template/partial/site-header }}
-
-<h1>{{ siteName }}</h1>
-
-<h2>Home template.</h2>
-
-<p>Calling the "test" method: <strong>{{ test }}</strong></p>
-
-<p>Template name: {{ templateName }}</p>
-
-{{> boilerplate/template/partial/site-footer }}
